backend/requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14,<3.16
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<0.13
psycopg[binary]>=3.1,<3.3
djangorestframework-simplejwt>=5.3,<6.0